[-h] Usage: 	/Users/slf/bin/bugtools/bugcli --help 	/Users/slf/bin/bugtools/bugcli close [--time=delta] [<id>] <comment> [file,file,...] 	/Users/slf/bin/bugtools/bugcli comment [--time=delta] [<id>] <comment> 	/Users/slf/bin/bugtools/bugcli commit [--time=delta] [<id>] <comment> [file,file,...] 	/Users/slf/bin/bugtools/bugcli create <product> <component> <summary> <description> [<assign to> [ <milestone> [ <blocksid> [ <blocksid> ... ] ] ] ] ] 	/Users/slf/bin/bugtools/bugcli fixed [--time=delta] [<id>] <comment> 	/Users/slf/bin/bugtools/bugcli full [<id>] 	/Users/slf/bin/bugtools/bugcli info [<id>] 	/Users/slf/bin/bugtools/bugcli list [--products | [ <product> [--components | <component>] [--bugstatus=<bugstatus>] ] ] 	/Users/slf/bin/bugtools/bugcli listbugs [bugid,bugid,...] 	/Users/slf/bin/bugtools/bugcli login <bugs login> 	/Users/slf/bin/bugtools/bugcli summary [<id>] 	/Users/slf/bin/bugtools/bugcli version <product> <version>
	Global options

		--no-id-update:   Does not update the current id

close:    Adds the comment to the bug, optionally adjusting time, and commits the code and comment to the SVN repo and marks the bug as fixed
comment:  Adds the comment to the bug, optionally adjusting time
commit:   Adds the comment to the bug, optionally adjusting time, and commits the code and comment to the SVN repo
create:   Create a new bug in the database
fixed:    Marks the bug as fixed, with the comment, optionally adjusting time
full:     Prints the full information of a bug
info:     Prints the product, component and summary of a bug
list:     List products, components or bugs
listbugs: List products like list but only by bugidlogin:    Creates the neccesary cookie files for future operations
summary:  Prints the summary of a bug
version:  Register a new version for the product

If no id is given, the id of the last succesfull command will be used.
Warning: The commit and close commands does not support files with spaces or other whitespace in the name. - 

git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@89387 b87a6194-a221-0410-ad6f-3b8ec35bba13