[21080] US #931: re-run file based job
Generate bit set to represent key indices for failed records if the
re-run should include failed records only. This bit set should then be
used to skip successful records.
JobEntity then needs to keep a counter of skipped items so that
DataPartitioner.drainItems can skip to the correct number if
partitioning is restarted.

git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@113530 b87a6194-a221-0410-ad6f-3b8ec35bba13