[21080] US #931: only include items in filter
If a bit set is generated with indices of items to be included then a
filter is made based on that set. Otherwise a filter always returning
true is used.
There is a problem of keeping track of item ids between the parent job
and the rerun job since the bit set is generated with values from the
parent job so these values must be used when filtering, but the indices
for items in the new job cannot be non-contiguous so new values must be
generated for them. And because the index set in the bit set relates to
both chunk id and item id we must also keep track of a virtual chunk id.

git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@113535 b87a6194-a221-0410-ad6f-3b8ec35bba13