Find sink in target flow store
Instead of creating a sink based on one in the source flow store, look
for one in the target flow store based on a name given as an argument
and report an error if it isn't found. This way we avoid jobs stuck in
delivery because they haven't got a sink to deliver to.