Establish a checkbox on the Jobs View page for disabling autorefresh
It has been investigated whether it is possible to either refresh a single
cell in the Jobs Table, OR assure, that a refresh does not remove focus,
and does not cause the display to be forced to display the top
lines only. A solution to that was not found.
Therefore, the disabling of the autorefresh was implemented instead.