Add clear log functionality
It is now posible to clear the log lines shown in
the ui. It is also possible to see the history of log lines
(will only be cleared on a browser refresh).
In time we might need to keep the log only for x amount of time
or until x size.