Show item diagnostics for partitioning phase
Up until now the "display diagnostic" functionality in the GUI assumed
that all diagnostics where contained in the state part of an item,
which is not the case for errors that are non-fatal at the job level,
in which case the diagnostics are kept in the phase specific part of the
item. Current workaround is to append diagnostics content to the item
data when fetched causing it to be shown inside the 'Input post' tab
instead of in the separate tabs otherwise used for fatal errors. These
separate tabs are not applicable in this case since their visibility is
determined prior to fetching the phase specific item data.