Use rawrepo version 1.8.
There was a bug/inconvenience in the agencyFor function in rawrepo-access, which meant in some rare cases a record for another library was wrongfully returned.
In order to get this fix rawrepo-access version 1.8 has to be used. Version 1.8 is instantiated a bit different that the previous version.

AgencySearchOrder has been replaced by a more elaborate RelationHint, which now requires a link to OpenAgency. Note that the url for AgencySearchOrder and OpenAgency is the same.