Add submitter scope to dependency tracking
The transition from json to arrays of hashes removed the submitter from
the match keys thus introducing the risk of database deadlocks since
chunks could now be made to wait for other chunks newer than themselves
when crossing submitter boundaries.

The addition of submitter scope eliminates the risk of deadlocks since
the job queue guarantees that jobs for the same sink and submitter are
partitioned sequentially and in order.