Don't fail on unmatched wiremock requests
Instruct WireMockRule to ignore unmatched requests. Otherwise the test
updateMarcXchange_wsFrameWorkThrows_throws fails with a
VerificationException when wiremock is updated to version 2.12.0.

But it seems weird to test that running without a stub results in a 404
error since this really only tests that wiremock is active but not
returning any data for the request, so perhaps the test is meant to test
something other than what it really tests.