Remove unnecessary throw of ConversionException
Only in case someone has manipulated the job-store database by hand
would the listJobs request return zero hits, and even then the situation
would eventually be handled by the DMQ functionality in the job-store.
Therefore it is deemed an unnecessary complexity to handle this case,
which would also require a failed chunk item to be produced.