Ensure arquillian test run against latest version
The bump from 5.191 to 5.193 caused one of the arquillian based test
suites to fail during embedded server startup. The actual cause of this
is the misconfiguration of the microservice-pom parent. The jersey
version used by payara is 2.29 and not 2.27. Overwriting this locally
until microservice-pom gets fixed enables the tests to run again.

Auditors: atm, jbr, hwha