Ensure HoldingsItemsConnector is closed
The IMS configurations need to do a Solr holdings lookup using the
HoldingsItemsConnector, which uses a solrj.HttpSolrClient which again
uses Apache HttpClient behind the covers. And since the Apache
HttpClient starts an IdleConnectionEvictor thread each time it is
instantiated, we leaked threads in the tens of thousands, because we
never deallocated the HoldingsItemsConnector resources after each
harvest operation.

Auditors: atm, jbr, hwha