Mitigate race condition during startup
Due to a thread safety issue in the Web application ClassLoader (reported to
and acknowledged by Payara support) when there are other EJB bean startup
initializations combined with a timer, it is possible that the EJB timer fires
(according to the defined schedule) before the deployment is finished.

Adding FlowStoreServiceConnectorBean to DependsOn will not make the
problem go away, but hopefully it will now fail during initialization
and thus make the service fail its health check.

Auditors: atm, jbr, hwha