Refactore to avoid extra rethrow of webservice exception
The reason for this refactoring is that we need the chunk to be able to log chunk- and jobid
when having a webservice exception. At the same time, we do want all exceptions caught and
reported. I prefer reporing unmarshall exceptions separately since we actually have an idea
of what is wrong at this stage.