Inject download url for dmat content
instead of using the configured baseurl value for the dmat connector.

This allows us to use a different url for downloads and harvesting items,
this might be needed if we need to use an alias for dmat towards end-users.
Also it gives us freedom to select the best path for the download endpoint
in the dmat service.